Small responsive fix, renaming home to register
